Split payment models into abstract and concrete
